Learning Py3 and making pricelists
At a current state it does the following:
1. Sorts the zoomos.csv in a way that allows to compare seller and supplier offers.
2. Tries to invent logic to compare thule.xls and zoomos.csv for the purpose of the above.

The main problem now is to reliably compare the .xls and .csv file for "Thule" positions.

In the future we're going to introduce user input and class methods, that will allow to address certain variable values and iterate through the whole zoomos. csv file at least.